* Changed db key from 'unsigned char*' to 'const char*'. * Defined out the dump function in util.h/c.
Notes: saw some suspicious code in npc.c so kept the 'const char*' propagation to a minimum for now

git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@10193 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/ers.c
@@ -349,7 +349,7 @@ static void ers_obj_destroy(ERS self)
  * @see #ers_root
  * @see #ers_num
  */
-ERS ers_new(size_t size)
+ERS ers_new(uint32 size)
 {
 	ERS_impl obj;
 	uint32 i;
